Scalar fields & fields value DAO in dedicated namespace
I made a first version of this patch that was very dumb.
I was just converting _ into \ with appropriate rename/move
to folders. However this lead to headaches for both PHPStorm
and phpunit (so to me) because we ended to have stuff like
FloatDao that where in 2 different namepaces Field/FloatDao
and Field/Value/FloatDao. It was barely readable and tbh
worst than the _ based namespace.

This second attempt tries to be a bit more smart by grouping
together things that "belongs" to the same field. I made it
for scalar fields only so far as a demo (and because it's less
invasive than for lists).

The mass rename .class.php -> .php was made for all because it
was easier to automate.

I had to regenerate the psalm baseline from scratch because of all
the moves.

Part of request #17148 Tracker move to PSR4 😱

Change-Id: Idae50bd60993a23653c8f298441ac4c6d51c8b1a