Mirror milestone also copies status with mapping
Part of story #16125 [release management] create a new Aggregator
Release

How to test:
- Set up multi projects
- Now that it is copied, "Status" field can be required in tracker
structures. Previously if it was required, it would fail the validation.
- In the Aggregator project, create a new Release and choose a Status
field.
- Mirror releases should be created in all the Contributor projects.
Each mirror release should have the same label for Status. The Status
value is mapped by duck-typing (if it has the same label, regardless of
casing, it is the same).

No functional change on "Move artifact" status mapping.

Change-Id: I4074a1728bc1d61b33c224d2c1311794e78b679e