Display background image in project (dashboard & admin only)
Run forgeupgrade, then hack the database to inject in project_background
table one of the two backgrounds provided by this changeset:
* beach
* bridge

Then load the project dashboard and the project administration: the
background should be displayed.
Other pages in the project should not have the background (for now).

Background images are taken from Unsplash[0] that allows us to use them
freely[1]:
> Unsplash grants you an irrevocable, nonexclusive, worldwide copyright
> license to download, copy, modify, distribute, perform, and use photos
> from Unsplash for free, including for commercial purposes, without
> permission from or attributing the photographer or Unsplash. This
> license does not include the right to compile photos from Unsplash to
> replicate a similar or competing service.

Images are in large format (2400px wide) and cropped to ≃850px height so
that header can be displayed on most pages without taking too much
bandwidth. Furthermore webp format is used instead of original jpg to
reduce the size of the asset.

Dominant color of the image is extracted[2] and displayed while the image
is loading, in order to allow a not too harsh transition.

Background is composed of 3 layers:

    Semi transparent opacifier
    so that the image integrates +--+
    with theme variants             |
                                    |
              +---------------------v-+
              |·   ·   ·   ·   ·   ·  |
           +--+  ·   ·   ·   ·   ·   ·|
           |  |·   ·   ·   ·   ·   ·  |
        +--+  |  ·   ·   ·   ·   ·   ·|
        |##|  |·   ·   ·   ·   ·   ·  |
        |##|  |  ·   ·   ·   ·   ·   ·|
        |##|  |·   ·   ·   ·   ·   ·  |
        |##|  |  ·   ·   ·   ·   ·   ·|
        |##|  +-------------------+---+
        |##|                      |
        |##+----------------+---^-+
        |###################|   |
        +^------------------+   |
         |                      |
         |       The image +----+
         |
         |
         +-+ Dominant color of the image
         * displayed while loading image
         * if image is not big enough,
           it is displayed as fallback

Note: We need to adjust FP tabs and feedback_* style so that they look
the same than on BP pages.

Note: While searching a place to put the background images, I found an
unused image image-homepage-tuleap.jpg. It has then been removed,
alongside the corresponding css style in src/themes/common/css/style.scss.

Part of story #16211 Set a header background image

[0] https://unsplash.com
[1] https://unsplash.com/license
[2] Ex of online tool: https://www.imgonline.com.ua/eng/get-dominant-colors-result.php

Change-Id: Ia71baa5c870f99acd0f9fc14a7715965ca9ab5b0