select/unselect multiple values
part of: story #16645 have searchable (multi) select boxes with (static binding, no dependencies)

Please note, unlike Select2, I have chosen not to sort out
alphabetically the list of the selected values each time a new value is
selected. To me it is useless and a bit unsettling. We'll see later with
UXD if he decides to keep the same bahavior. Also please note that the
selection of items through keyboard is not implemented yet.

How to test:
- In an empty multi-list-picker, select several values
--> the placeholder disappears
--> a "remove all values" button is added in the selection_element
--> The values appear selected in the dropdown
--> Selected values are displayed in badges
--> They have a "remove" button

- Unselect a value by clicking its remove button
--> it is unselected
--> the dropdown opens

- Unselect all values by clicking on the "remove all values" button
--> all values are unselected
--> the dropdown opens

- Select several values, then hit the backspace key
--> the last selected value is unselected
--> it's template is displayed in the search input
--> the dropdown opens
--> the list is immediately filtered

Change-Id: Ie157dfcb28c62d4a521c8be84d9861f90fd0205d