Project admin can choose the background (Frontend)
Bootstrap the background selection by displaying available backgrounds
in project admin.

A preview of each background is displayed as well as its label and the
author.

While the backgrouds can be selected, the settings cannot be saved yet,
it will be done in a dedicated changeset.

Part of story #16211 Set a header background image

Change-Id: I01142341ee0d5ff8a860e3c534614e8dfaec742a