Transform hardly visible browser deprecation banner into a popup
The popup is shown once every 4 days so it is seen by the users at
least once per work week.

Transforming the banner into a modal also means it will have less
interactions with the rest of Tuleap and will not need a refactoring for
upcoming features such the site wide banner.

Closes request #17155: Announce end of support of Internet Explorer

Change-Id: Ie7298434d002e8f0319ddcc93745aba6d2b2eef2