Display background on every pages of the project
We are changing the strategy here: instead of having to inject the
ProjectBackgroundConfiguration on every controllers and every front
scripts in project, we now do this only once in the Theme while
displaying the header. This allows us to not repeat ourselves and miss a
particular script.

Now every pages in the project displays the background. Most of the
times it will be only part of the background since we need to adjust
styles, but since it is hidden behind a feature flag, it is considered
acceptable. Styling will be adjusted as things progress in dedicated
changesets.

As before, if feature flag is off, then only project dashboards have the
background.

In order to test this, you need to have a project with a background:
PUT /projects/:id/header_background
{ "identifier": "beach" }

Part of story #16211 Set a header background image

Change-Id: I359bd6047f3b05c088eeb40c71d2630bdbcae5be