Refactoring: do not repeat no-repeat
No functional changes.

Information about how the background should be displayed belongs to the
background definition. We should not have to choose if the image should
repeat itself or to choose its size every time we use the background.

This also prepares the field for SVG background.

Part of story #16211: Set a header background image

Change-Id: I18d74810f2227ce966c55dc4b6d54f8f06cc96c5