Refactoring: minor fixes
part of story #16125 [release management] create a new Aggregator
Release

No functional change expected.
The "Status" namespace should be only for Status semantic-related
classes. The field id parameters in the exception was probably
mass-renamed accidentally, so this patch changes that.
SynchronizedFields does not need to re-expose TimeframeFields'
getStartDateField since we need to get the TimeframeFields to access the
EndPeriod anyway. We should keep the two APIs apart.

Change-Id: Ie94ce27b7addd3f9da661a331ca5a86e08ea5d8f