Use fa-lock-open instead of fa-unlock for project privacy
Two recent events decreased the readability of the priacy icon:
* Introduction of background image in project header
  => with a background, there is no distinctive color between a "public"
  and a "private" project.
* Upgrade to Font Awesome 5
  => the `fa fa-unlock` is not the same icon than `fas fa-unlock`

The association  makes it hard to distinquish a "public" and "private"
icon.

Font Awesome 5 `fa-lock-open` matches the old `fa-unlock` so we are now
using it for project privacy icon (furthermore it matches the
fa-tlp-unlock-plus-r custom icon).

Usage of `fa-unlock` outside of project privacy icon is ignored by this
patch.

Part of story #16211: Set a header background image

Change-Id: Ibe186b5333405750ffb4a7075dbb28af8d509375