Try to not load large header background image on small screen
Not loading a large image on a small screen can help saving resources.
It also make conceivable to ship images that could be displayed on 4k
screen (or at least a WQHD one) without making sacrifices.

This contribution ships images with a 1680px width to cover what appears
to be the vast majority of the users according to multiple statistics
sources. It looks like if we want to cover most of our users including
the ones running high end hardware we would need to ship 3 different
sizes: 1680px > 1920px > 2560px || 3840px.

Images have been converted with the following command (with exception of
green-grass thta has been cropped at 0:0):
 $> find . -name '*.jpg' -exec bash -c 'cwebp -preset photo -mt -m 6 -pass 10 -af -crop 360 0 1680 850 $1 -o ${1%.jpg}-1680px.webp' -- {} \;

Part of story #16211: Set a header background image

Change-Id: Id9992679d5b9c6ef83e7cf466ed151cad4d778fe