request #17533 Workflow doesn't applied in modal creation
How to reproduce:
	- Have a required selectbox in a tracker
	- Add a workflow on this field (Created -> Open, Open -> Closed)
    	- Open a modal with this tracker
        	- You can choose between "", 'Closed', 'Open' instead of only 'Open'

This issue was introduce in story #16125 [0]
Previously, the 'is_used' field in tracker workflow in Rest API was a string, and modal use it.
Now it's a boolean stored in db (0 or 1), and it's an integer that's displayed in Rest API, and the modal can't use is anymore.

[0]  git #tuleap/stable/1e59908135dbd7d072c22aa89c07de2f14dc42cb

Change-Id: Id19e56e8ed6d7d40c091fcb47bbbe326f2e63867