request #17537: Remove hardcoded ellipsis of cell titles
The titles are cut as soon as they are longer than 50 characters
regardless of the available size to display the title.

This approach cause more harm than good:
* it can makes harder for users to find the document they need to interact with
* it does not solve the core issue: since the ellipsis truncation is static and does
not take into account the actual available display size the truncated string can still
be longer than what can be displayed...

Letting the title wrap on a second line is not that bad here (it's still
an edge case anyway) and it's already what the breadcrumd does.

Change-Id: I13825a98c1a1b0f6fa33d350f880377887128f92