request #17874: Too much items in project admin menu
The tabs of project admin contain too much item, and decrease UX on this
part:
* Details
* Members
* Groups
* Permissions
* Services
* Labels
* References
* Catégories
* Data
* Banner
* Background
* Project links
* OAuth2 apps
* Project ownership

This is too much, most of the items are rarely used (maybe once during
the first configuration of the project).

The menu should focus on important or regularly used items (adding a new
member, checking permissions, …)

The proposal is to move other items in a "hamburger" menu like it has
been done in tracker administration:

The menu will then look like:
* Members
* Groups
* Permissions
* Details
* Data
* (Others)

Furthermore, since managing membership is considered as the most used
item, we propose that it becomes the default one.

Change-Id: Id4decbf2730b6506f1cb9ead9c58201ceff20152