Display Subversion Core in Plugin
This introduce the ability for SVN Plugin to deal with the Subversion
repositories historically managed by Core "Subversion" service.

At this stage it's purely a display patch, the goal is to introduce
the underlying objects and interfaces to distinguish the repository
types.

The patch does a bit of clean-up in term of types to leverage psalm
assistance while developing the features.

Part of request #17893 Phase-out core "Subversion" service

Change-Id: Ib9b272a335d6a0402c28194f51532d7aaf80dcb5