Use the list picker in the angular modal
Part of: story #16646 have searchable (multi) select boxes with (static binding, no dependencies) in modal

How to test:
 - Go to a page where the angularjs modal is used. e.g: kanban, planning v2 etc...
 - Open the modal
=> All the select are replaced by the new list picker
=> There is no fucntionnal change at the artifact submission.

Note:
We could not use the angularjs depsite the reduced template implementation.
The main reason is the "$element" is not provided in a component and only provided if we use a directive.

Change-Id: I2bd4e75f00eed4c171651642656f6b3171c7d470