Display project background in Git and pull request
The administration sections will be dealt with in a separate
contribution. Those sections still being under the FlamingParrot theme.

You might also notice the transparency of some tlp elements is not great
when the project background is used (e.g. the tlp-button-bar-checkbox
or the disabled button elements). Since changing this as site-wide
consequences this will also be done in a separate contributions.

In order to test you need:
* the feature flag feature_flag_project_background activated
* a project with a background
* another project without background to compare differences

Part of story #17551: apply background image on all project pages

Change-Id: I1f67725c4497c71ade1f023422fdfcbf128e2153