Create REST GET projects/:id/gitlab_repositories
This is part of story #17178 reference my gitlab repositories in my tuleap project

A new REST route GET projects/:id/gitlab_repositories is created.
It retrieves the Gitlab repositories integrated into a sepcific project.

To test it, you have to:

* Uninstall and reinstall the Gitlab plugin
* Enable this plugin
* Add data by hand in the plugin_gitlab_repositories SQL table.

This route is paginated, but not in SQL. This will be done in a dedicated commit.

Change-Id: I85e45d8e248cb71635893c4625015323c73e7e3f