Wrong menu position when project has background
Default tlp-tabs have a 2px border at the bottom to indicate the active
tab. The submenu is therefore moved 2px under.

However, for project tabs, when the project has a background image, the
2px border is at the top. Therefore the menu should not be moved 2px
under, else there is a gap between the trigger and the menu and the menu
can be hidden before you have a chance to reach it, which is quite
frustrating.

Identified by request #17874: Too much items in project admin menu
(but issue is for every tlp-tab-menu in project administration)

Change-Id: I0579169be58631052a9bc85ba99f7d76c19a9a3d