CVS header reorganisation
In order to better display the project background image in CVS service,
the toolbar is removed in favor of tabs.

In the process, some simplifications have been made:
* Remove "CVS" term from menu items: we already know that we are in
cvs service thanks to the new title or the new breadcrumb item.
* Remove "help" memu item: to be consistent with other services
* Remove "My commits" menu item: it is a shortcut that targets CVS query
("Commits" tab) prefiltered. Since this feature is not available in
recent VCS implementations (Git/SVN plugins) we propose to 🔥 it.

Part of story #17551: apply background image on all project pages

Change-Id: I8a0b9a4cfdb3ddb4e20ad2984b503a57c1cbc62c