request #17909: Cannot display an artifact affected by only one workflow transition rule
Reproduction scenario:
 1. With the workflow in the simple mode, one (and only one) transition
    between two values
 2. Display an artifact with a value at the beginning of this transition

This issue is caused by the fact the TransitionExtractor class assume
that State::getTransitions() returns a list. This is in fact not always
true as one of the call stacks start in
TransitionFactory::getTransitions() where the transitions are indexed by
their internal IDs. In any cases the code should avoid such assumption
when not working with an explicit list<T> array.

Change-Id: Ibc224b739936f4baee8711553ca5aa47f472dc16