Remove colored top border of kanban columns when a project background is used
The colored border is not really needed as there is a clear cut between
the two content and dependending on the theme and the background image
the mix between the two is not always great.

The colored top border is kept when the kanban is used in a widget.
Typography changes of the filter label and text info also have been
excluded when the kanban is used in a widget as it makes the message
hard to read.

In order to test you need:
* the feature flag feature_flag_project_background activated
* a project with a background
* another project without background to compare differences

Part of story #17551: apply background image on all project pages

Change-Id: I4ee363ca6cef168fb2c2223366fc95cf0c9fa128