Tests should propagate the PHP version used
This is needed when the PHP version used is not the current default one.
The change in SystemEventProcessor_Root class has been done to propagate the
environment variables, the goal is to let the php-launcher.sh script
have access to the PHP env variable when processing the system events.

It should be noted that anyway the Docker image running the REST tests
will need to have both PHP versions with all the extensions due to the
way the create_test_env plugin works.

Part of request #17910: Compatibility with Composer 2

Change-Id: Ia925e48b93e0a6b9a23f2a43ab8f2bc5c3e6c900