Fix GitLab trademark casing
GitLab is always written with an uppercase L and it's how the trademark
is registered [0]. We should stick to this casing.

Note the README.mkd file has been removed instead of fixed. It contained
the same content than the description of the plugin without being
internationalized so it brought little value.

Part of story #17178: reference my gitlab repositories in my tuleap project

[0] https://about.gitlab.com/handbook/marketing/growth-marketing/brand-and-digital-design/brand-guidelines/#trademark

Change-Id: I09c543cad22c7bf1f0931cd81c2d3fe844b92cab