SVN Core ➡️ Plugin migration with data in place
Initially the migration of svn core to plugin involved move of
data from svnroot to svnplugin prefix (with new name, path & all).

The previous commit demonstrate that we can perfectly have two kind
of storage (svnroot or svn_plugin) without touching the URLs (the
most important stuff because it's involve in svn:externals).

This commit changes the way migration is done to actually "delegate"
the core repository management to the plugin. That basically means
that Core should no longer interfer in delegated repositories business.

The initial conditional display of "CoreRepository" in svn plugin
view is removed because the migration is the action to promote.

Please note that the action is more or less reversible in db (for
tests) by removing the repository with `is_core = 1` for the project.
This might evolve toward a real feature (for test) in a future commit.

Part of request #17893 Phase-out core "Subversion" service

Change-Id: Ib354c8070b6732532bee627380d80d676e3c99c0