Do not enable processing of external entities when validating the schema of XML files
This change is needed because PHP 8.0 deprecates usage of \libxml_disable_entity_loader() [0].
It also a good thing from a security PoV because it is one less spot
where things can go bad.

To achieve that, the RNG files are flatten before runtime so there is no
need to include files when a XML files needs to be verified against a
schema.
A new Docker image is used to do the RNC to RNG transformation. It replaces
the old one that was quite outdated and dealing with a lot of other
things.

Part of request #17931: Unit tests should pass with PHP 8.0

[0] https://github.com/php/php-src/blob/php-8.0.0RC3/UPGRADING#L873

Change-Id: I360c67ccef75980acb02a74cda60ebddc5c2be18