Remove \XML_Security::loadFile() and \XML_Security::loadString()
PHP 8 deprecates the \libxml_disable_entity_loader() function [0].
Calling XML parsing methods and functions can be considered safe by default,
there is no more need to wrap the calls inside another class. So to be
sure new usages of this class are not introduced inside the codebase we
reduce the exposed API to the bare minimum.

Part of request #17931: Unit tests should pass with PHP 8.0

[0] https://github.com/php/php-src/blob/php-8.0.0RC3/UPGRADING#L873

Change-Id: Ic7bc9b45b6c7335205bf503945c026aed22295b8