Display project banner
Issue PUT /bannner to create a platform banner, or use site admin
utilities.

Once the platform banner is set, it is displayed on every pages of the
platform.

If user is logged in, then the banner can be hidden. Once hidden, the
banner won't be automatically displayed on page load unless it is
updated (or removed/recreated). The user can still have a look at the
message though, thanks to the bullhorn floating button.

Note: If both platform and project banners are displayed, and the
platform has "warning" importance, then the background color of the two
banners will be the same orange, without any distinction. This issue
will be managed in a dedicated commit.

This replaces the legacy motd.

There are many situtations that should be checked. You can mix any
combinations of these:
- Pages with or without a sidebar (project, siteadmin)
- Pages with FlamingParrot or BurningParrot
- Project banner not set, set and displayed, or set and hidden
- Platform banner not set, set and displayed, unclamped, or set and hidden
- To scroll or not to scroll
- Choice of browser

Please note that on applications that compute their height, the presence
of the platform banner may force the viewport to scroll and hide some
content under the fold. It is not ideal and we should see how we can fix
this situation. However I suggest that we consider this acceptable:
- The platform banner is intended to be closed as soon as the message is
read. If user continues to browse with the platform banner displayed on
every pages, then having to scroll should not be a real problem.
- The hidden stuff under the fold are controls (TTM buttons, add inplace
for kanban, …) that are not available for users that cannot hide the
banner (they are anonymous).
Given that the impact is low, we can live with it for this commit.

Clamping (like for project banner), if needed, will be done in a
dedicated changeset.

Part of story #14670: display a site wide banner (motd)

Change-Id: I453f22114da1035dad9cb5e6b2dfdba3751ed6c5