Create REST endpoint DELETE gitlab_repositories/:id
This is part of story #17178 reference my gitlab repositories in my tuleap project

The REST route DELETE gitlab_repositories/:id is now available. Project
id must be provided as an URL parameter and is mandatory.

The integration is removed if user is git administrator of the project. If the last
integration is removed, all the GitLab repository data is also removed.

Errors are thrown if repository does not exist or if repository is not
integrated into provided project.

Change-Id: I62bc062dee74b71c4a5a9d0a6554fe6176cd7de9