request #17958: Legacy logo is always displayed except on freshly installed instances
git #tuleap/stable/bc4d2ca1bfb646188ac81ee5e9ccced8dfc1be02 have
rewritten all PNG files which cause the legacy organization logo
to always be detected as customized.
Since fresh installations might have the new version of the logo we are
forced to change the check a bit so we can deal with changes in the logo.

Change-Id: Ieb6cffb9a99f8f7bc77d47a8218a38fa061cf06d