Create REST endpoint to integrate a GitLab repository
This is part of story #17178 reference my gitlab repositories in my tuleap project

A new REST endpoint is now available: POST gitlab_repository.

The expected JSON body content must be like:
{
  "project_id": 122,
  "gitlab_server_url" : "https://example.com",
  "gitlab_user_api_token" : "my_token",
  "gitlab_internal_id" : 145896
}

If this is the first time the GitLab repository is integrated, then the
whole intgeration process is done:

* Add data in repository table
* Add integration in project
* Create a secret for the repository
* Add a new post push webhook in the GitLab project.

If this GitLab repository is already integrated in at least one Tuleap
project, then only a new entry in repository_project table is added.

Change-Id: Ia195fbd6844235e67a022c4f410f86771b09e457