request #17985 Scrollbar in Switch to modal should not take focus
On Firefox, an element with a scrollbar due to `overflow-y:auto` takes focus to actionate the scrollbar.
In the Switch to modal, this prevents the user from tabing from the `input#switch-to-filter` directly to the first list item.
When going down using the scrollbar, the next tabbed element will however be the first list item.

This prevents the scrollbar from taking focus.

Change-Id: I117e4fe268b94d120110cdf857859625a1fc2e19