Open modal to create an activity in a kanban
…instead of leaving the current page and losing the context.

Given a kanban is based on a tracker (e.g Activity) that is promoted in
the «+ New» dropdown,
When user goes to the given kanban
And select the "Add new Activity" in the "+ New» dropdown,
Then the modal to create the activity is displayed just like if the user
would have clicked on the primary button "Add new Activity" above the
kanban.

Note: in the rare case where the tracker is not configured enough to
have the primary button "Add new activity" (e.g lack of semantic), then
the old behavior is kept (redirection to the tracker service with a
possible loss of context). The criticity of this case is considered low
enough to not have to handle it right now.

Part of story #16206: « + New » button, better contextualisation

Change-Id: I5e165907578578a4c677ee50352c52f02e9516c1