Auto-promote current tracker in +New dropdown on kanban
Given my kanban is based on Activities tracker
And Activities is not promoted to +New dropdown
When I go to kanban,
Then "New Activities" is displayed in the +New dropdown, in the current
section labelled "Kanban".

If Activities tracker is already promoted, then "New Activities" will
be displayed only once (like it is done in a tracker).

If the user cannot add artifact directly to the kanban
(i.e. there is no primary button to add activities), then the link is
not added to the +New dropdown.

Part of story #16206: « + New » button, better contextualisation

Change-Id: Idffb31be375cd918e1139822d2b8f43cf9d0f8d7