request #18360: Keep formatting when switching between text and html formats
Go to a tracker artifact, edit a text field that is containing some html
formatting (list, headings, …). Switch the format to text => the textarea
does not contain html tag soup.

Switch back to html => the WYSIWYG editor does not contain the text on
one line only like it was before. Thanks to the intermediate markdown
format, the formatting is kept.

This is useful when you start to right your answer in text format and
realize that you will need more formatting. Before that, switching to
html format was producing garbage, you had to copy the text beforehand.

Please note that this contribution does not fix all issues. For example
if you dragndrop an image in HTML format and switch to text format, then
the text will contain a reference to the image that cannot be exploited,
just like before when the text was containing <img> tag. This issue will
be managed properly with the introduction of the upcoming Markdown
editor.

Change-Id: I80f25efc28b36377c7da1e51f058743bdeacfb15