Stop reacting to an update of an agile dashboard planning
As the planning configuration of a "moderately scaled agile" dashboard
will be not be accessible when the project is either a team or a program
project. As a result the event PlanningUpdateEvent and the associated
code in the Scaled Agile plugin is soon to be dead code and safely be
removed.

Part of story #17879: add Teams to Programs

Change-Id: Ibc4b9b6daff911f66df8ef61313dcf60724f2010