Allow creation of log files on the "backend SVN" part of the ditributed Tuleap integration tests
Since 63e6b66faaca6e924f2a7847848b45b51a16416a, Tuleap tries to create
log files in a way that make sure they are going to be writable.
This is an issue in the distributed Tuleap integration tests because the
image used to run the "backend SVN" part is minimal. Some parts that
are usually managed by the packaging are not done at all making
impossible to write the Tuleap log files.

Part of request #18331: Local log files should be created with the appropriate ownership

Change-Id: Ia56008ac3c416391b7b2f7b828ae35e0306b4e19