Better contextualisation of +New for submilestones
Currently in planning v2 interface, when a user is able to add a
submilestone (e.g a release in the top backlog or a sprint in a release)
via a dedicated button on top of the list of submilestones. This button
opens a modal.

The proposal of this contribution is to auto promote the submilestone
tracker in the +New dropdown and to open the same modal than the button
when user clicks on it (like it has been done for kanban interface).

Part of story #16206: « + New » button, better contextualisation

Change-Id: I47ef57379f1066484f72693772d287466a36f217