Allow tracker creation/deletion alongside tracker promotion
Since request #18330, there is a new interface in tracker global
administration dedicated to trackers management. As of today it only
handles promotion of trackers in +New dropdown, but trackers admin
should be able to delete or create trackers from this interface as well.

The aim of this commit is to introduce the tracker creation/deletion in the
tracker global administration. Removal of tracker deletion in the
trackers homepage will be done in a dedicated commit, in order to not
clutter the review.

Part of request #18352: Creation or deletion should be available in tracker global admin interface

Change-Id: I0f1322ea86afa0b0d42c3daa39e63bcc43fccb84