Display form error when the List Picker value is not valid in the angular modal
Part of story #16646 have searchable (multi) select boxes with (static binding, no dependencies) in modal

The error is displayed when a List Picker field is required and if there is no value selected.

NB: It also displays the form error for the "classic" select multiple but not for the single select.
NB2: This is also a part of request #13467 Artifact-modal allows empty required text fields when format is HTML and list fields
It seems this contribution resolves the list fields issue part of this request. If the List Picker is on, the user cannot submit the artifact.

Change-Id: Ib891a4f0bc64c31bda559980ad4a8278e0f65a0e