request #17989: Forum service cannot be used with Amazon SES
The mails sent by the forum service are sent without a To header.
While those sort of mails are valid, Amazon SES rejects them.
Amazon SES is likely not the only service having issue with those sort of mails
(and anyway it might affects a non negligible amount of users given the market share of Amazon).

All the emails sent by Tuleap without a To header now gets one set to
the value of sys_noreply. The same strategy was already used by the
massmail tool available in the site administration, it is now
generalized to all the mails.

Change-Id: I321646e1927596a41ccc272dd4b5826b9d43a6ff