Better contextualisation of parent trackers in AD
Same as git #stable/tuleap/afffc1d6c17066cf9ae3246f15876ec3c5f05d87 [0], but
for parent trackers that are proposed in the Add item dropdown in pv2
(section "parents")  =>
on every panes of the milestone, such parent trackers are auto promoted
in the +New menu. And on pv2, this opens the modal instead of going to
the tracker page.

Part of story #16206: « + New » button, better contextualisation

[0] https://gerrit.tuleap.net/c/tuleap/+/20884

Change-Id: Ibf42bbb23a741780c6fbee8676c02fab8ef40f8d