Mailman administration is not usable
Since b767bc5c127d08c58a1a468fe5a602c125c3a176 a default
Content-Security-Policy blocking everything is set by default.
For most Tuleap pages this is fine because one is set, in the case of
mailman it is not because we end up blocking all the interactions.

Part of request #17967: Deploy a useful content security policy

Change-Id: I2647c6da1b6ab66566cf2a0ba6f7440c80453a10