Makes the displayed keyboard input in Shortcut Help Modal easily readable
Part of story #18347 : Improve the shortcuts Help modal

Adds the possibility to make displayed keyboard input different from the keyboard input used to trigger the shortcut action.
For example, `CTRL+up` is replaced by `CTRL+↑` which can be easier to read and understand.

Changes the `kbd` elements font to monospace font to ensure all letters can be easily distinguished.
For example, `l` should not be mistaken for an `i` or an `1`.

Change-Id: Idb10392b0b6ea3d89eb203fefa47c8a761b64475