Parse TULEAP-{:id} references in commit messages
This is part of story #17180 take into account gitlab commits

The commits provided by the post push webhook content are now parsed.
TULEAP-{:id} references (case insensitive with id as an integer) are now
extracted. They are not saved in DB for now, only logged in the dedicated
log file.

Change-Id: I7835454dc08b5975083395053f32f45e04d5e8b2