request #18376: Prevent PHP and nginx versions to be sent in response headers
While this has very little to no benefit from a security point of view
(please just update your softwares instead of trying to hide publicly known
information), automated tools flag this as a very serious issue. It is
easier to drop the versions than to argue with automated tools.

To test, you need to redeploy the nginx configuration. Please note that
on a dev setup there is a reverse proxy which also add its own "Server"
header, so you need to query directly the server without going through
the reverse proxy to see the change.

Change-Id: Ia317706449eb1cb6af6caf90ef6763384f5204ed