Block hierarchy management in the tracker admininistration for trackers involved in a scaled agile hierarchy
Note that the tracker administration does not manage very well the
possibility of having hierarchy defined accross multiple projects at the
moment. To test, either do it in isolation or workaround to have a
scaled agile hierarchy in the same project.

Part of story #17882: define cross project hierarchy between Features and User Stories

Change-Id: I4112778e63d1890b1594afc856ed01b9bed0b530