Refactoring: extract agiledashboard redirection in dedicated object
In order to prepare the field for future contributions, we should have a
net safety by extracting logic in dedicated object with better test
coverage.

Redirection was introduced in planning view v1, but mostly disappeared
with introduction of pv2 and its artifact modal. Redirection was used to
be able to create/edit artifact in tracker and being redirected to the
agiledashboard in the right pane, with autolinking.

Today, this redirection mechanism is only used in details (overview)
pane (and in legacy cardwall, but its irrelevant for the current
contribution). I'm not sure if all redirect features are still used, but
I blindly extracted the whole stuff in order to focus on the next
addition: redirection from the +New dropdown that will use more or less
the same mechanism.

No functional changes are expected.

Part of story #16206: « + New » button, better contextualisation

Change-Id: I95e2b5076f5548dcab16411acad898219d960c07